Addressed CR comment (separated ObjectInspector wrapper into its own file)
